fix(tool): keep csr generator files under the max-lines limit
- Split tools/csr-generator/core/csr.ts (487 raw / ~425 effective
  lines) into core/csr.ts (composition + Subject/SAN builders),
  core/keys.ts (key generation, import, PKCS#8 parsing), and
  core/errors.ts (shared error type). Each stays well under the
  300-line oxlint max-lines budget.
- Extract the key-source / algorithm / hash / curve / import-PEM
  section of components/csr-options-card.tsx into a new
  components/csr-key-source-fields.tsx so the parent card drops
  from 349 lines to 134.
- Fix smart-quote JSON parse errors in messages/de.json,
  messages/pl.json, and messages/zh-CN.json that broke the
  Astro/Vite JSON loader.
